# docs/swagger.yaml
basePath: /
definitions:
...
  endpoint.createPatientRequest:
    properties:
      address:
        example: 123 Main St
        type: string
      age:
        example: 30
        type: integer
      email:
        example: john@example.com
        type: string
      full_name:
        example: John Doe
        type: string
      gender:
        example: Male
        type: string
      health_history:
        example:
        - Diabetes
        - Hypertension
        items:
          type: string
        type: array
      job:
        example: Engineer
        type: string
      password:
        example: password123
        type: string
      patient_code:
        example: J001
        type: string
      phone_number:
        example:
        - "081234567890"
        - "081234567891"
        items:
          type: string
        type: array
      surgery_history:
        example: Appendectomy 2020
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  model.Patient:
    description: Patient information
    properties:
      address:
        example: 123 Main St
        type: string
      age:
        example: 30
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        example: john@example.com
        type: string
      full_name:
        example: John Doe
        type: string
      gender:
        example: Male
        type: string
      health_history:
        example: Diabetes,Hypertension
        type: string
      id:
        example: 1
        type: integer
      job:
        example: Engineer
        type: string
      password:
        example: hashed_password
        type: string
      patient_code:
        example: J001
        type: string
      phone_number:
        example: "081234567890"
        type: string
      surgery_history:
        example: Appendectomy 2020
        type: string
      updatedAt:
        type: string
    type: object
...
  util.APIResponse:
    properties:
      data: {}
      error:
        type: string
      msg:
        type: string
      success:
        type: boolean
    type: object
host: localhost:19091
info:
  contact:
    email: support@ariebrainware.com
    name: Arie Brainware
  description: |-
    REST API for managing patient data, treatments, and therapy sessions
    This API provides endpoints for patient management, disease tracking, treatment records, and therapist management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: LTT Backend API
  version: "1.0"
paths:
...
  /patient:
    get:
      consumes:
      - application/json
      description: Get a paginated list of patients with optional filtering
      parameters:
      - description: Limit number of results
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Search keyword for patient name, code, address, or phone
        in: query
        name: keyword
        type: string
      - description: Filter by date range (last_2_days, last_3_months, last_6_months)
        in: query
        name: group_by_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Patients retrieved
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/util.APIResponse'
      security:
      - BearerAuth: []
      - SessionToken: []
      summary: List all patients
      tags:
      - Patient
    post:
      consumes:
      - application/json
      description: Register a new patient (public endpoint - no authentication required)
      parameters:
      - description: Patient information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/endpoint.createPatientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Patient created
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Invalid request or patient already exists
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Create a new patient
      tags:
      - Patient
  /patient/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a patient by ID
      parameters:
      - description: Patient ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Patient deleted
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Patient not found
          schema:
            $ref: '#/definitions/util.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/util.APIResponse'
      security:
      - BearerAuth: []
      - SessionToken: []
      summary: Delete a patient
      tags:
      - Patient
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific patient
      parameters:
      - description: Patient ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Patient retrieved
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Patient'
              type: object
        "400":
          description: Patient not found
          schema:
            $ref: '#/definitions/util.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/util.APIResponse'
      security:
      - BearerAuth: []
      - SessionToken: []
      summary: Get patient information
      tags:
      - Patient
    patch:
      consumes:
      - application/json
      description: Update an existing patient's information
      parameters:
      - description: Patient ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated patient information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Patient'
      produces:
      - application/json
      responses:
        "200":
          description: Patient updated
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Patient'
              type: object
        "400":
          description: Invalid request or patient not found
          schema:
            $ref: '#/definitions/util.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/util.APIResponse'
      security:
      - BearerAuth: []
      - SessionToken: []
      summary: Update patient information
      tags:
      - Patient
...
  /token/validate:
    get:
      consumes:
      - application/json
      description: Validate if the session token is valid and not expired
      produces:
      - application/json
      responses:
        "200":
          description: Valid session token
          schema:
            $ref: '#/definitions/util.APIResponse'
        "401":
          description: Invalid or expired session token
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/util.APIResponse'
      security:
      - BearerAuth: []
      - SessionToken: []
      summary: Validate session token
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  SessionToken:
    description: Session token for authenticated requests
    in: header
    name: session-token
    type: apiKey
swagger: "2.0"