# endpoint/patient.go
package endpoint

import (
	"fmt"
	"strconv"
	"strings"
	"time"

	"github.com/ariebrainware/basis-data-ltt/middleware"
	"github.com/ariebrainware/basis-data-ltt/model"
	"github.com/ariebrainware/basis-data-ltt/util"
	"github.com/gin-gonic/gin"
	"gorm.io/gorm"
)

func parseQueryParams(c *gin.Context) (int, int, int, string, string) {
	limit, _ := strconv.Atoi(c.Query("limit"))
	offset, _ := strconv.Atoi(c.Query("offset"))
	therapistID, _ := strconv.Atoi(c.Query("therapist_id"))
	keyword := c.Query("keyword")
	groupByDate := c.Query("group_by_date")
	return limit, offset, therapistID, keyword, groupByDate
}

func applyGroupByDateFilter(query *gorm.DB, groupByDate string) *gorm.DB {
	switch groupByDate {
	case "last_2_days":
		query = query.Where("created_at >= ?", time.Now().AddDate(0, 0, -2))
	case "last_3_months":
		query = query.Where("created_at >= ?", time.Now().AddDate(0, -3, 0))
	case "last_6_months":
		query = query.Where("created_at >= ?", time.Now().AddDate(0, -6, 0))
	}
	return query
}

func fetchPatients(db *gorm.DB, limit, offset, therapistID int, keyword, groupByDate string) ([]model.Patient, int64, error) {
	var patients []model.Patient
	var totalPatient int64

	query := db.Offset(offset).Order("created_at DESC")
	if limit > 0 {
		query = query.Limit(limit)
	}
	if offset > 0 {
		query = query.Offset(offset)
	}
	if keyword != "" {
		query = query.Where("full_name LIKE ? OR patient_code LIKE ? OR address LIKE ? OR phone_number LIKE ?", "%"+keyword+"%", "%"+keyword+"%", "%"+keyword+"%", "%"+keyword+"%")
	}
	query = applyGroupByDateFilter(query, groupByDate)

	if err := query.Find(&patients).Error; err != nil {
		return nil, 0, err
	}

	db.Model(&model.Patient{}).Count(&totalPatient)
	return patients, totalPatient, nil
}

// ListPatients godoc
// @Summary      List all patients
// @Description  Get a paginated list of patients with optional filtering
// @Tags         Patient
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Security     SessionToken
// @Param        limit query int false "Limit number of results"
// @Param        offset query int false "Offset for pagination"
// @Param        keyword query string false "Search keyword for patient name, code, address, or phone"
// @Param        group_by_date query string false "Filter by date range (last_2_days, last_3_months, last_6_months)"
// @Success      200 {object} util.APIResponse{data=object} "Patients retrieved"
// @Failure      401 {object} util.APIResponse "Unauthorized"
// @Failure      500 {object} util.APIResponse "Server error"
// @Router       /patient [get]
func ListPatients(c *gin.Context) {
	limit, offset, therapistID, keyword, groupByDate := parseQueryParams(c)

	db := middleware.GetDB(c)
	if db == nil {
		util.CallServerError(c, util.APIErrorParams{
			Msg: "Database connection not available",
			Err: fmt.Errorf("db is nil"),
		})
		return
	}

	patients, totalPatient, err := fetchPatients(db, limit, offset, therapistID, keyword, groupByDate)
	if err != nil {
		util.CallServerError(c, util.APIErrorParams{
			Msg: "Failed to retrieve patients",
			Err: err,
		})
		return
	}

	util.CallSuccessOK(c, util.APISuccessParams{
		Msg:  "Patients retrieved",
		Data: map[string]interface{}{"total": totalPatient, "total_fetched": len(patients), "patients": patients},
	})
}

type createPatientRequest struct {
	FullName       string   `json:"full_name" example:"John Doe"`
	Gender         string   `json:"gender" example:"Male"`
	Age            int      `json:"age" example:"30"`
	Job            string   `json:"job" example:"Engineer"`
	Address        string   `json:"address" example:"123 Main St"`
	PhoneNumber    []string `json:"phone_number" example:"081234567890,081234567891"`
	HealthHistory  []string `json:"health_history" example:"Diabetes,Hypertension"`
	SurgeryHistory string   `json:"surgery_history" example:"Appendectomy 2020"`
	PatientCode    string   `json:"patient_code" example:"J001"`
	Password       string   `json:"password,omitempty" example:"password123"`
	Email          string   `json:"email,omitempty" example:"john@example.com"`
}

// CreatePatient godoc
// @Summary      Create a new patient
// @Description  Register a new patient (public endpoint - no authentication required)
// @Tags         Patient
// @Accept       json
// @Produce      json
// @Param        request body createPatientRequest true "Patient information"
// @Success      200 {object} util.APIResponse "Patient created"
// @Failure      400 {object} util.APIResponse "Invalid request or patient already exists"
// @Failure      500 {object} util.APIResponse "Server error"
// @Router       /patient [post]
func CreatePatient(c *gin.Context) {
	patientRequest := createPatientRequest{}

	err := c.ShouldBindJSON(&patientRequest)
	if err != nil {
		util.CallUserError(c, util.APIErrorParams{
			Msg: "Invalid request body",
			Err: err,
		})
		return
	}
	if patientRequest.FullName == "" || len(patientRequest.PhoneNumber) == 0 {
		util.CallUserError(c, util.APIErrorParams{
			Msg: "Patient payload is empty or missing required fields",
			Err: fmt.Errorf("invalid payload"),
		})
		return
	}

	db := middleware.GetDB(c)
	if db == nil {
		util.CallServerError(c, util.APIErrorParams{
			Msg: "Database connection not available",
			Err: fmt.Errorf("db is nil"),
		})
		return
	}

	var existingPatient model.Patient
	var existingUser model.User

	err = db.Transaction(func(tx *gorm.DB) error {
		// Check if username and phone already registered
		if err := tx.Where("full_name = ? AND (phone_number = ? OR phone_number IN ?)", patientRequest.FullName, strings.Join(patientRequest.PhoneNumber, ","), patientRequest.PhoneNumber).First(&existingPatient).Error; err != nil {
			if err != gorm.ErrRecordNotFound {
				return fmt.Errorf("database error: %v", err)
			} // else continue to create new patient
		}

		// determine the patient code by fullname initials + incremented number
		initials := getInitials(patientRequest.FullName)

		// generate patient.patient_code based on patient_codes table
		var patientCodeTable model.PatientCode
		err := tx.Order("id DESC").Where("alphabet = ?", initials).First(&patientCodeTable).Error
		if err != nil {
			if err == gorm.ErrRecordNotFound {
				return fmt.Errorf("patient code not found")
			} else {
				return err
			}
		}

		var patientCode string
		if patientRequest.PatientCode != "" {
			patientCode = patientRequest.PatientCode
		} else {
			newNumber := patientCodeTable.Number + 1
			patientCode = fmt.Sprintf("%s%d", initials, newNumber)
			if err := tx.Where("alphabet = ?", initials).Updates(&model.PatientCode{
				Number:   newNumber,
				Alphabet: initials,
				Code:     patientCode,
			}).Error; err != nil {
				return err
			}
		}

		// Check if patients.patient_code already registered
		if err := tx.Where("patient_code = ?", patientCode).First(&existingPatient).Error; err == nil {
			return fmt.Errorf("patient_code already registered")
		}

		if patientRequest.Email != "" && patientRequest.Email != "-" {
			if patientRequest.Password != "" {
				// Check if users already registered
				if err := tx.Where("email = ?", patientRequest.Email).First(&existingUser).Error; err == nil {
					return fmt.Errorf("email already registered")
				}
				// Create user
				if err := tx.Create(&model.User{
					Name:     patientRequest.FullName,
					Email:    patientRequest.Email,
					Password: util.HashPassword(patientRequest.Password),
					RoleID:   2,
				}).Error; err != nil {
					return err
				}
			}
		}

		if err := tx.Create(&model.Patient{
			FullName:       patientRequest.FullName,
			Gender:         patientRequest.Gender,
			Age:            patientRequest.Age,
			Job:            patientRequest.Job,
			Address:        patientRequest.Address,
			PhoneNumber:    strings.Join(patientRequest.PhoneNumber, ","),
			PatientCode:    patientCode,
			HealthHistory:  strings.Join(patientRequest.HealthHistory, ","),
			SurgeryHistory: patientRequest.SurgeryHistory,
			Email:          patientRequest.Email,
			Password:       util.HashPassword(patientRequest.Password),
		}).Error; err != nil {
			return err
		}
		return nil
	})

	if err != nil {
		util.CallServerError(c, util.APIErrorParams{
			Msg: "Failed to create patient",
			Err: err,
		})
		return
	}

	util.CallSuccessOK(c, util.APISuccessParams{
		Msg:  "Patient created",
		Data: nil,
	})
}

// UpdatePatient godoc
// @Summary      Update patient information
// @Description  Update an existing patient's information
// @Tags         Patient
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Security     SessionToken
// @Param        id path string true "Patient ID"
// @Param        request body model.Patient true "Updated patient information"
// @Success      200 {object} util.APIResponse{data=model.Patient} "Patient updated"
// @Failure      400 {object} util.APIResponse "Invalid request or patient not found"
// @Failure      401 {object} util.APIResponse "Unauthorized"
// @Failure      500 {object} util.APIResponse "Server error"
// @Router       /patient/{id} [patch]
func UpdatePatient(c *gin.Context) {
	id := c.Param("id")
	if id == "" {
		util.CallUserError(c, util.APIErrorParams{
			Msg: "Missing patient ID",
			Err: fmt.Errorf("patient ID is required"),
		})
		return
	}

	patient := model.Patient{}
	if err := c.ShouldBindJSON(&patient); err != nil {
		util.CallUserError(c, util.APIErrorParams{
			Msg: "Invalid request body",
			Err: err,
		})
		return
	}

	db := middleware.GetDB(c)
	if db == nil {
		util.CallServerError(c, util.APIErrorParams{
			Msg: "Database connection not available",
			Err: fmt.Errorf("db is nil"),
		})
		return
	}

	var existingPatient model.Patient
	if err := db.First(&existingPatient, id).Error; err != nil {
		util.CallUserError(c, util.APIErrorParams{
			Msg: "Patient not found",
			Err: err,
		})
		return
	}

	if err := db.Model(&existingPatient).Updates(patient).Error; err != nil {
		util.CallServerError(c, util.APIErrorParams{
			Msg: "Failed to update patient",
			Err: err,
		})
		return
	}

	util.CallSuccessOK(c, util.APISuccessParams{
		Msg:  "Patient updated",
		Data: existingPatient,
	})
}

func getInitials(fullName string) string {
	words := strings.Fields(fullName)
	initials := ""
	if len(words) > 0 && len(words[0]) > 0 {
		initials = strings.ToUpper(string(words[0][0]))
	}
	return initials
}

func getPatientByID(c *gin.Context, db *gorm.DB) (string, model.Patient, error) {
	id := c.Param("id")
	if id == "" {
		util.CallUserError(c, util.APIErrorParams{
			Msg: "Missing patient ID",
			Err: fmt.Errorf("patient ID is required"),
		})
		return "", model.Patient{}, fmt.Errorf("patient ID is required")
	}

	var patient model.Patient
	if err := db.First(&patient, id).Error; err != nil {
		util.CallUserError(c, util.APIErrorParams{
			Msg: "Patient not found",
			Err: err,
		})
		return "", model.Patient{}, err
	}

	return id, patient, nil
}

// DeletePatient godoc
// @Summary      Delete a patient
// @Description  Soft delete a patient by ID
// @Tags         Patient
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Security     SessionToken
// @Param        id path string true "Patient ID"
// @Success      200 {object} util.APIResponse "Patient deleted"
// @Failure      400 {object} util.APIResponse "Patient not found"
// @Failure      401 {object} util.APIResponse "Unauthorized"
// @Failure      500 {object} util.APIResponse "Server error"
// @Router       /patient/{id} [delete]
func DeletePatient(c *gin.Context) {
	db := middleware.GetDB(c)
	if db == nil {
		util.CallServerError(c, util.APIErrorParams{
			Msg: "Database connection not available",
			Err: fmt.Errorf("db is nil"),
		})
		return
	}

	_, patient, err := getPatientByID(c, db)
	if err != nil {
		return
	}

	if err := db.Delete(&patient).Error; err != nil {
		util.CallServerError(c, util.APIErrorParams{
			Msg: "Failed to delete patient",
			Err: err,
		})
		return
	}

	util.CallSuccessOK(c, util.APISuccessParams{
		Msg: "Patient deleted",
	})
}

// GetPatientInfo godoc
// @Summary      Get patient information
// @Description  Get detailed information about a specific patient
// @Tags         Patient
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Security     SessionToken
// @Param        id path string true "Patient ID"
// @Success      200 {object} util.APIResponse{data=model.Patient} "Patient retrieved"
// @Failure      400 {object} util.APIResponse "Patient not found"
// @Failure      401 {object} util.APIResponse "Unauthorized"
// @Failure      500 {object} util.APIResponse "Server error"
// @Router       /patient/{id} [get]
func GetPatientInfo(c *gin.Context) {
	db := middleware.GetDB(c)
	if db == nil {
		util.CallServerError(c, util.APIErrorParams{
			Msg: "Database connection not available",
			Err: fmt.Errorf("db is nil"),
		})
		return
	}

	_, patient, err := getPatientByID(c, db)
	if err != nil {
		return
	}

	util.CallSuccessOK(c, util.APISuccessParams{
		Msg:  "Patient retrieved",
		Data: patient,
	})
}